A LOT of changes

+++ IDQ.Domain/Models/consumoProductoModel.cs
@@ -5,13 +5,13 @@ namespace IDQ.Domain.Models
     public class consumoProductoModel : Base.ModelBase
     {
         #region Private
-        int _Cantidad; Double _Precio; fechaModel _Fecha; productoModel _Producto;
+        int _Cantidad; Decimal _Precio; fechaModel _Fecha; productoModel _Producto;
         #endregion // Private
 
         #region Public
         public int Cantidad { get => _Cantidad; set { if (SetProperty(ref _Cantidad, value)) { OnPropertyChanged(); } } }
 
-        public Double Precio { get => _Precio; set { if (SetProperty(ref _Precio, Math.Round(value, 2))) { OnPropertyChanged(); } } }
+        public Decimal Precio { get => _Precio; set { if (SetProperty(ref _Precio, Math.Round(value, 2))) { OnPropertyChanged(); } } }
 
         public int FechaID { get; set; }
         public virtual fechaModel Fecha { get => _Fecha; set { if (SetProperty(ref _Fecha, value)) { OnPropertyChanged(); } } }
